Static analysis fixes
#jira UE-102442, UE-102438
#rb martin.ridgers

[CL 14651299 by Johan Berg in ue5-main branch]

+++ Engine/Source/Runtime/Core/Private/ProfilingDebugging/Windows/CallstackTrace.cpp
@@ -60,7 +60,7 @@ static uint32 AddressToId(UPTRINT Address)
 
 static UPTRINT IdToAddress(uint32 Id)
 {
-	return UPTRINT(Id << 16);
+	return static_cast<uint32>(UPTRINT(Id) << 16);
 }
 
 struct FIdPredicate
